requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
bidict==0.23.1
boto3==1.34.54
botocore==1.34.54
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepspeed==0.13.4
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1698714947081/work
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
hjson==3.1.0
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==7.0.1
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1655912915439/work
jmespath==1.0.1
joblib==1.3.2
langcodes==3.3.0
llvmlite==0.41.1
locate==1.1.1
Markdown==3.5.2
MarkupSafe==2.0.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
msgpack==1.0.8
multidict==6.0.5
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1680692919326/work
ninja==1.11.1.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==4.25.3
psutil==5.9.8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1684180539862/work
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
torch_cluster==1.6.3
torch_geometric==2.5.0
torch_scatter==2.1.2
torch_sparse==0.6.18
torch_spline_conv==1.2.2
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
transformers==4.38.2
triton==2.2.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
tzdata==2024.1
urllib3==1.26.18
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.1
wordfreq==3.1.1
yarl==1.9.4
zipp==3.17.0
